s19b: split jeep_liq.c at its constructor, rodata to jeep2/jeep_liq/spark2_m
The s19b_spark2_m_800D8724..800DA0B4 block (Liquid's attack handlers)
lived in jeep_liq.c, but the original TU boundary is right after the
jeep_liq.c constructor (800D8670): the byte after its "liqiud" string
is uninitialized in the original (0xAF in s19b, 0xAC in s19br), which
only happens between objects. The block therefore moves to the top of
spark2_m.c, the next object in link order, and both files share the
actor's Work through the new jeep_liq.h (spark2_m's own work struct
is renamed SparkWork).

With that boundary, the .rdata that overlay5.c was holding can go
where the compiler will put it once the switches are matched:
"jeep2.c" into jeep2.c; the 800D7468 jump table, the 800D8250
constant, the 8420 strings and "jeep_liq.c"/"liqiud" into jeep_liq.c;
the 800D9AA8 jump table and the 800D9EC0 constant into spark2_m.c.
overlay5.c (both overlays) keeps only its .data words.

um/s19b.bin and um/s19br.bin are re-extracted: the pad after "liqiud"
is now a real linker gap, so each sequence gains one byte.
