+++ source/overlays/s19b/overlay5.c
@@ -23,24 +23,3 @@ int s19b_dword_800C3AC8 = 0x00140014;
 int s19b_dword_800C3ACC = 0x000C0010;
 int s19b_dword_800C3AD0 = 0x98660005;
 int s19b_dword_800C3AD4 = 0x00000000;
-
-const char s19b_aJeepc_800DDE44[] = "jeep2.c";
-const int s19b_dword_800DDE4C = 0x800D74B0;
-const int s19b_dword_800DDE50 = 0x800D74F8;
-const int s19b_dword_800DDE54 = 0x800D753C;
-const int s19b_dword_800DDE58 = 0x800D75B0;
-const int s19b_dword_800DDE5C = 0x800D768C;
-const char s19b_dword_800DDE60[] = {'^', 0x1, 0xcf, 0x3};
-const char s19b_dword_800DDE64[] = {0xa2, 0xfe, 0x0, 0x0};
-const char s19b_aLiquid_800DDE68[] = "liquid";
-const char s19b_aLqdb_800DDE70[] = "lqd_19b";
-const char s19b_aFamas_800DDE78[] = "famas";
-const char s19b_aJeepliqc_800DDE80[] = "jeep_liq.c";
-const char s19b_dword_800DDE8C[] = {'l', 'i', 'q', 'i', 'u', 'd', 0x0, 0xaf};
-const int s19b_dword_800DDE94 = 0x800D9AE0;
-const int s19b_dword_800DDE98 = 0x800D9AF0;
-const int s19b_dword_800DDE9C = 0x800D9B00;
-const int s19b_dword_800DDEA0 = 0x800D9B10;
-const int s19b_dword_800DDEA4 = 0x800D9B20;
-const char s19b_dword_800DDEA8[] = {0x0, 0x0, 0x0, 0x0};
-const char s19b_dword_800DDEAC[] = {'2', 0x2, 0x0, 0x0};
