+++ source/overlays/s19b/takabe/jeep2.c
@@ -41,7 +41,7 @@ void s19b_jeep2_800D6AD8(Jeep2Work *work)
 
 extern void s19b_jeep2_800D667C(Jeep2Work *work);
 extern int  s19b_jeep2_800D6B18(Jeep2Work *work, int arg0, int arg1);
-extern const char s19b_aJeepc_800DDE44[];
+const char s19b_aJeepc_800DDE44[] = "jeep2.c";
 
 void *s19b_jeep2_800D6F24(int arg0, int arg1)
 {
