+++ source/overlays/s19b/takabe/jeep_liq/jeep_liq.c
@@ -1,107 +1,10 @@
+#include "jeep_liq.h"
+
+#include <rand.h>
 #include "common.h"
 #include "game/game.h"
 #include "linkvar.h"
 #include "psxdefs.h"
-#include <rand.h>
-
-typedef struct _Work
-{
-    GV_ACT          actor;        /* 0x000 */
-    CONTROL        *control;      /* 0x020 */
-    DG_PRIM        *prim;         /* 0x024 */
-    CONTROL         ctrl;         /* 0x028 */
-    OBJECT          obj;          /* 0x0A4 */
-    MOTION_CONTROL  m_ctrl;       /* 0x188 */
-    /* m_segs1[17] @ 0x1D8 - poked as scalars below by the setters */
-    char            pad_seg1[0x1F0 - 0x188 - sizeof(MOTION_CONTROL)];
-    int             f1F0;         /* 0x1F0 */
-    char            pad_360[0x360 - 0x1F4];
-    int             field_360;    /* 0x360 */
-    char            pad_364[0x378 - 0x360 - sizeof(int)];
-    short           field_378;    /* 0x378 */
-    char            pad_37A[0x37C - 0x378 - sizeof(short)];
-    short           field_37C;    /* 0x37C */
-    char            pad_37E[0x390 - 0x37C - sizeof(short)];
-    SVECTOR         field_390;    /* 0x390 */
-    char            pad_398[0x3A0 - 0x390 - sizeof(SVECTOR)];
-    int             field_3A0;    /* 0x3A0 */
-    int             field_3A4;    /* 0x3A4 */
-    int             field_3A8;    /* 0x3A8 */
-    int             field_3AC;    /* 0x3AC */
-    int             field_3B0;    /* 0x3B0 */
-    int             field_3B4;    /* 0x3B4 */
-    int             field_3B8;    /* 0x3B8 */
-    char            pad_3BC[0x3C8 - 0x3B8 - sizeof(int)];
-    int             field_3C8;    /* 0x3C8 */
-    int             field_3CC;    /* 0x3CC */
-    int             field_3D0;    /* 0x3D0 */
-    int             field_3D4;
-    int             field_3D8;
-    int             field_3DC;
-    int             field_3E0;
-    int             field_3E4;
-    int             field_3E8;
-    int             field_3EC;    /* 0x3EC */
-    int             field_3F0;    /* 0x3F0 */
-    char            pad_3F4[0x3F8 - 0x3F0 - sizeof(int)];
-    int             field_3F8;    /* 0x3F8 */
-    char            pad_segs2[0x43C - 0x3F8 - sizeof(int)];
-    MOTION_SEGMENT  m_segs2[17];  /* 0x43C */
-    SVECTOR         svecs1[16];   /* 0x6A0 */
-    SVECTOR         svecs2[16];   /* 0x720 */
-    SVECTOR         sv_7A0;       /* 0x7A0 */
-    SVECTOR         sv_7A8;       /* 0x7A8 */
-    char            pad_7B0[0x7B8 - 0x7A8 - sizeof(SVECTOR)];
-    SVECTOR         sv_7B8;       /* 0x7B8 */
-    MATRIX          mtx[2];       /* 0x7C0 */
-    OBJECT          obj2;         /* 0x800 */
-    TARGET         *f8E4;         /* 0x8E4 */
-    HOMING         *homing;       /* 0x8E8 */
-    void           *f8EC;         /* 0x8EC */
-    char            pad_8F4[0x8F4 - 0x8EC - sizeof(void *)];
-    int             f8F4;         /* 0x8F4 */
-    char            pad_8FC[0x8FC - 0x8F4 - sizeof(int)];
-    int             f8FC;         /* 0x8FC */
-    int             f900;         /* 0x900 */
-    int             f904;         /* 0x904 */
-    char            pad_90C[0x90C - 0x904 - sizeof(int)];
-    int             f90C;         /* 0x90C */
-    int             f910;         /* 0x910 */
-    int             f914;         /* 0x914 */
-    int             f918;         /* 0x918 */
-    char            pad_920[0x920 - 0x918 - sizeof(int)];
-    int             f920;         /* 0x920 */
-    char            pad_930[0x930 - 0x920 - sizeof(int)];
-    int             f930;         /* 0x930 */
-    int             f934;         /* 0x934 */
-    char            pad_93C[0x93C - 0x934 - sizeof(int)];
-    int             f93C;         /* 0x93C */
-    int             f940;         /* 0x940 */
-    int            *f944;         /* 0x944 */
-} Work;
-
-typedef struct _JEEP_SYSTEM_S
-{
-    char     pad1[0x4];
-    CONTROL *control;
-    char     pad2[0x10];
-    SVECTOR  pos;
-    char     pad3a[0x24 - 0x18 - sizeof(SVECTOR)];
-    short    field_24;
-    char     pad3b[0x30 - 0x24 - sizeof(short)];
-    int      field_30;
-    char     pad3c[0x40 - 0x30 - sizeof(int)];
-    int      field_40;
-    int      field_44;
-    char     pad3d[0x54 - 0x44 - sizeof(int)];
-    int      field_54;
-    char     pad5c[0x5C - 0x54 - sizeof(int)];
-    int      field_5C;
-    char     pad60[0x138 - 0x5C - sizeof(int)];
-    SVECTOR  field_138;
-} JEEP_SYSTEM_S;
-
-extern JEEP_SYSTEM_S Takabe_JeepSystem;
 
 extern int  s19b_dword_800C3AA0;
 extern int  s19b_dword_800C3AA8;
@@ -136,9 +39,7 @@ extern void  s19b_jeep_liq_800D7C0C(Work *work);
 extern void  s19b_jeep_liq_800D8014(Work *work, int arg1);
 extern void  s19b_jeep_liq_800D8118(Work *work);
 extern void  s19b_jeep_gls_800CEC24(int arg0, SVECTOR *out);
-extern void  s19b_spark2_m_800D964C(Work *work);
 extern void  sna_act_helper2_helper2_80033054(int id, SVECTOR *vec);
-extern const char s19b_dword_800DDE60[];
 extern int   s19b_jeep_gls_800CEDFC(int arg0, int arg1);
 
 void s19b_jeep_liq_800D6FB8(Work *work)
@@ -163,6 +64,13 @@ void s19b_jeep_liq_800D6FB8(Work *work)
     work->field_3C8++;
     work->field_3AC = work->field_3A8;
 }
+/* jump table of s19b_jeep_liq_800D7468 */
+const int s19b_dword_800DDE4C = 0x800D74B0;
+const int s19b_dword_800DDE50 = 0x800D74F8;
+const int s19b_dword_800DDE54 = 0x800D753C;
+const int s19b_dword_800DDE58 = 0x800D75B0;
+const int s19b_dword_800DDE5C = 0x800D768C;
+
 void s19b_jeep_liq_800D7114(Work *work)
 {
     int state;
@@ -661,13 +569,15 @@ void s19b_jeep_liq_800D81A8(Work *work)
         work->field_3D0 = (int)s19b_jeep_liq_800D7114;
     }
 }
+const SVECTOR s19b_dword_800DDE60 = {0x15E, 0x3CF, -0x15E, 0};
+
 void s19b_jeep_liq_800D8250(Work *work)
 {
     OBJECT  *body = &work->obj;
     CONTROL *ctl = &work->ctrl;
     DG_OBJS *objs;
     SVECTOR *jpos = &Takabe_JeepSystem.field_138;
-    SVECTOR  pos = *(SVECTOR *)&s19b_dword_800DDE60;
+    SVECTOR  pos = s19b_dword_800DDE60;
 
     GM_ActMotion(body);
     ctl->rot = ctl->turn = DG_ZeroVector;
@@ -696,6 +606,10 @@ void s19b_jeep_liq_800D8250(Work *work)
         body->objs->flag &= ~0x80;
     }
 }
+const char s19b_aLiquid_800DDE68[] = "liquid";
+const char s19b_aLqdb_800DDE70[] = "lqd_19b";
+const char s19b_aFamas_800DDE78[] = "famas";
+
 #pragma INCLUDE_ASM("asm/overlays/s19b/s19b_jeep_liq_800D8420.s")
 void s19b_spark2_m_800D8620(Work *work)
 {
@@ -707,8 +621,8 @@ void s19b_spark2_m_800D8620(Work *work)
 }
 extern void s19b_jeep_liq_800D8250(Work *work);
 extern int  s19b_jeep_liq_800D8420(Work *work, int name);
-extern const char s19b_aJeepliqc_800DDE80[];
-extern const char s19b_dword_800DDE8C[];
+const char s19b_aJeepliqc_800DDE80[] = "jeep_liq.c";
+const char s19b_dword_800DDE8C[] = "liqiud";
 
 void *s19b_spark2_m_800D8670(CONTROL *arg0, DG_PRIM *arg1, int *arg2)
 {
@@ -728,862 +642,4 @@ void *s19b_spark2_m_800D8670(CONTROL *arg0, DG_PRIM *arg1, int *arg2)
         }
     }
     return work;
-}
-
-void s19b_spark2_m_800D8724(Work *work, int arg1, int arg2)
-{
-    MATRIX  m;
-    DG_OBJ *obj = &work->obj.objs->objs[arg1];
-
-    DG_SetPos(&obj->world);
-    DG_MovePos((SVECTOR *)&s19b_dword_800C3AA0);
-    DG_RotatePos((SVECTOR *)&s19b_dword_800C3AA8);
-    ReadRotMatrix(&m);
-    NewJeepBlood(&m, arg2, &obj->world);
-}
-
-extern void s19b_spark2_m_800D9390(Work *work, int mode);
-extern void s19b_spark2_m_800D9434(Work *work, int mode);
-extern void s19b_spark2_m_800D94C8(Work *work, int mode);
-
-#pragma INCLUDE_ASM("asm/overlays/s19b/s19b_spark2_m_800D87A4.s")
-extern int s19b_spark2_m_800D87A4(Work *work);
-
-int s19b_spark2_m_800D88D8(Work *work)
-{
-    if (s19b_spark2_m_800D87A4(work) != 0)
-    {
-        work->f8E4->class = 1;
-        return 1;
-    }
-    return 0;
-}
-
-void s19b_spark2_m_800D8918(Work *work)
-{
-    TARGET *target = work->f8E4;
-    int     level  = GM_GameLevel;
-    int     vital;
-
-    if (level > 0)
-    {
-        vital = (level << 6) + 0xBF;
-    }
-    else
-    {
-        vital = 0xBF;
-    }
-
-    GM_SetTarget(target, 20, 2, (SVECTOR *)&s19b_dword_800C3AB0);
-    GM_SetPowerTarget(target, 1, -1, vital, 7, (SVECTOR *)&s19b_dword_800C3AB8);
-}
-
-extern void s19b_spark2_m_800D8BC8();
-extern void s19b_spark2_m_800D8CEC();
-extern void s19b_spark2_m_800D8E10();
-extern void s19b_spark2_m_800D90A8();
-extern void s19b_spark2_m_800D92C8();
-
-int s19b_spark2_m_800D899C(Work *work)
-{
-    int v = work->f900;
-
-    if (v & 0x1)
-    {
-        work->f8EC = (void *)s19b_spark2_m_800D8BC8;
-        work->f8F4 = 0;
-        work->ctrl.turn.vz = 0;
-        work->ctrl.turn.vx = 0;
-        return 1;
-    }
-    if (v & 0x2)
-    {
-        work->f8EC = (void *)s19b_spark2_m_800D8CEC;
-        work->f8F4 = 0;
-        work->ctrl.turn.vz = 0;
-        work->ctrl.turn.vx = 0;
-        return 1;
-    }
-    if (v & 0x4)
-    {
-        work->f8EC = (void *)s19b_spark2_m_800D8E10;
-        work->f8F4 = 0;
-        work->ctrl.turn.vz = 0;
-        work->ctrl.turn.vx = 0;
-        return 1;
-    }
-    if (v & 0x10)
-    {
-        work->f8EC = (void *)s19b_spark2_m_800D90A8;
-        work->f8F4 = 0;
-        work->ctrl.turn.vz = 0;
-        work->ctrl.turn.vx = 0;
-        return 1;
-    }
-    if (v & 0x04000000)
-    {
-        work->f8EC = (void *)s19b_spark2_m_800D92C8;
-        work->f8F4 = 0;
-        work->ctrl.turn.vz = 0;
-        work->ctrl.turn.vx = 0;
-        return 1;
-    }
-    return 0;
-}
-
-void s19b_spark2_m_800D8A48(Work *work)
-{
-    GV_NearExp4PV(&work->svecs2[2], &work->sv_7A0, 3);
-    GV_NearExp4PV(&work->svecs2[6], &work->sv_7A8, 3);
-}
-
-void s19b_spark2_m_800D8A88(Work *work)
-{
-    int v = work->f930 - work->control->rot.vy;
-    work->sv_7A0.vy = v;
-    work->sv_7A8.vy = v;
-}
-
-void s19b_spark2_m_800D8AAC(Work *work)
-{
-    int base = work->f930 + 0xCC0;
-    int v    = base - work->control->rot.vy;
-    work->sv_7A0.vy = v;
-    work->sv_7A8.vy = v;
-}
-
-void s19b_spark2_m_800D8ACC(Work *work)
-{
-    int base = work->f930 + 0x340;
-    int v    = base - work->control->rot.vy;
-    work->sv_7A0.vy = v;
-    work->sv_7A8.vy = v;
-}
-
-void s19b_spark2_m_800D8AEC(Work *work)
-{
-    MATRIX m;
-
-    DG_SetPos(&work->obj.objs->objs[4].world);
-    DG_MovePos((SVECTOR *)&s19b_dword_800C3AC0);
-    ReadRotMatrix(&m);
-    NewJeepBullet(&m, 2, 1, 0);
-    GM_SeSet(&work->ctrl.mov, 0x2E);
-    s19b_jblood_800C7FB8(&m);
-}
-
-void s19b_spark2_m_800D8B54(Work *work, int mode)
-{
-    if (mode == 0)
-    {
-        work->f8FC = 0;
-        GM_ConfigObjectAction((OBJECT *)&work->obj, 0, 0, 4);
-    }
-    if (s19b_spark2_m_800D88D8(work) == 0)
-    {
-        if (s19b_spark2_m_800D899C(work) == 0)
-        {
-            work->f8E4->class |= 0x14;
-        }
-    }
-}
-
-extern void s19b_spark2_m_800D8F34(Work *work, int mode);
-
-void s19b_spark2_m_800D8BC8(Work *work, int mode)
-{
-    int f900 = work->f900;
-
-    if (mode == 0)
-    {
-        work->f8FC = 1;
-        GM_ConfigObjectAction((OBJECT *)&work->obj, 1, 0, 4);
-    }
-    s19b_spark2_m_800D8A88(work);
-
-    if (!(f900 & 1))
-    {
-        work->f8EC = (void *)s19b_spark2_m_800D8B54;
-        work->f8F4 = 0;
-        work->ctrl.turn.vz = 0;
-        work->ctrl.turn.vx = 0;
-        work->sv_7A0 = DG_ZeroVector;
-        work->sv_7A8 = DG_ZeroVector;
-        return;
-    }
-
-    if (f900 & 8)
-    {
-        work->f8EC = (void *)s19b_spark2_m_800D8F34;
-        work->f8F4 = 0;
-        work->ctrl.turn.vz = 0;
-        work->ctrl.turn.vx = 0;
-        return;
-    }
-
-    if (s19b_spark2_m_800D88D8(work) != 0) return;
-    if (s19b_spark2_m_800D899C(work) != 0) return;
-    work->f8E4->class |= 0x14;
-}
-
-extern void s19b_spark2_m_800D8FB0(Work *work, int mode);
-
-void s19b_spark2_m_800D8CEC(Work *work, int mode)
-{
-    int f900 = work->f900;
-
-    if (mode == 0)
-    {
-        work->f8FC = 3;
-        GM_ConfigObjectAction((OBJECT *)&work->obj, 3, 0, 4);
-    }
-    s19b_spark2_m_800D8AAC(work);
-
-    if (!(f900 & 2))
-    {
-        work->f8EC = (void *)s19b_spark2_m_800D8B54;
-        work->f8F4 = 0;
-        work->ctrl.turn.vz = 0;
-        work->ctrl.turn.vx = 0;
-        work->sv_7A0 = DG_ZeroVector;
-        work->sv_7A8 = DG_ZeroVector;
-        return;
-    }
-
-    if (f900 & 8)
-    {
-        work->f8EC = (void *)s19b_spark2_m_800D8FB0;
-        work->f8F4 = 0;
-        work->ctrl.turn.vz = 0;
-        work->ctrl.turn.vx = 0;
-        return;
-    }
-
-    if (s19b_spark2_m_800D88D8(work) != 0) return;
-    if (s19b_spark2_m_800D899C(work) != 0) return;
-    work->f8E4->class |= 0x14;
-}
-
-extern void s19b_spark2_m_800D902C(Work *work, int mode);
-
-void s19b_spark2_m_800D8E10(Work *work, int mode)
-{
-    int f900 = work->f900;
-
-    if (mode == 0)
-    {
-        work->f8FC = 2;
-        GM_ConfigObjectAction((OBJECT *)&work->obj, 2, 0, 4);
-    }
-    s19b_spark2_m_800D8ACC(work);
-
-    if (!(f900 & 4))
-    {
-        work->f8EC = (void *)s19b_spark2_m_800D8B54;
-        work->f8F4 = 0;
-        work->ctrl.turn.vz = 0;
-        work->ctrl.turn.vx = 0;
-        work->sv_7A0 = DG_ZeroVector;
-        work->sv_7A8 = DG_ZeroVector;
-        return;
-    }
-
-    if (f900 & 8)
-    {
-        work->f8EC = (void *)s19b_spark2_m_800D902C;
-        work->f8F4 = 0;
-        work->ctrl.turn.vz = 0;
-        work->ctrl.turn.vx = 0;
-        return;
-    }
-
-    if (s19b_spark2_m_800D88D8(work) != 0) return;
-    if (s19b_spark2_m_800D899C(work) != 0) return;
-    work->f8E4->class |= 0x14;
-}
-
-void s19b_spark2_m_800D8F34(Work *work, int mode)
-{
-    if (mode == 0)
-    {
-        s19b_spark2_m_800D8AEC(work);
-    }
-    if (s19b_spark2_m_800D88D8(work) != 0)
-    {
-        return;
-    }
-    if (mode == 1)
-    {
-        work->f8EC = (void *)s19b_spark2_m_800D8BC8;
-        work->f8F4 = 0;
-        work->ctrl.turn.vz = 0;
-        work->ctrl.turn.vx = 0;
-    }
-    work->f8E4->class |= 0x14;
-}
-
-void s19b_spark2_m_800D8FB0(Work *work, int mode)
-{
-    if (mode == 0)
-    {
-        s19b_spark2_m_800D8AEC(work);
-    }
-    if (s19b_spark2_m_800D88D8(work) != 0)
-    {
-        return;
-    }
-    if (mode == 1)
-    {
-        work->f8EC = (void *)s19b_spark2_m_800D8CEC;
-        work->f8F4 = 0;
-        work->ctrl.turn.vz = 0;
-        work->ctrl.turn.vx = 0;
-    }
-    work->f8E4->class |= 0x14;
-}
-
-void s19b_spark2_m_800D902C(Work *work, int mode)
-{
-    if (mode == 0)
-    {
-        s19b_spark2_m_800D8AEC(work);
-    }
-    if (s19b_spark2_m_800D88D8(work) != 0)
-    {
-        return;
-    }
-    if (mode == 1)
-    {
-        work->f8EC = (void *)s19b_spark2_m_800D8E10;
-        work->f8F4 = 0;
-        work->ctrl.turn.vz = 0;
-        work->ctrl.turn.vx = 0;
-    }
-    work->f8E4->class |= 0x14;
-}
-
-extern void s19b_spark2_m_800D9148(Work *work, int mode);
-extern void s19b_spark2_m_800D91DC(Work *work, int mode);
-
-void s19b_spark2_m_800D90A8(Work *work, int mode)
-{
-    if (mode < 8)
-    {
-        work->f8E4->class |= 0x14;
-    }
-    if (s19b_spark2_m_800D88D8(work) != 0)
-    {
-        return;
-    }
-    if (mode == 0)
-    {
-        work->f8FC = 0x10;
-        GM_ConfigObjectAction((OBJECT *)&work->obj, 0x10, 0, 4);
-    }
-    if (work->obj.is_end != 0)
-    {
-        work->f8EC = (void *)s19b_spark2_m_800D9148;
-        work->f8F4 = 0;
-        work->ctrl.turn.vz = 0;
-        work->ctrl.turn.vx = 0;
-    }
-}
-
-void s19b_spark2_m_800D9148(Work *work, int mode)
-{
-    int f900 = work->f900;
-
-    if (!(f900 & 0x10))
-    {
-        work->f8EC = (void *)s19b_spark2_m_800D91DC;
-        work->f8F4 = 0;
-        work->ctrl.turn.vz = 0;
-        work->ctrl.turn.vx = 0;
-        return;
-    }
-    if (s19b_spark2_m_800D88D8(work) != 0) return;
-    if (mode == 0)
-    {
-        work->f8FC = 0x11;
-        GM_ConfigObjectAction((OBJECT *)&work->obj, 0x11, 0, 4);
-    }
-    if (mode >= 0x5B)
-    {
-        work->f8EC = (void *)s19b_spark2_m_800D91DC;
-        work->f8F4 = 0;
-        work->ctrl.turn.vz = 0;
-        work->ctrl.turn.vx = 0;
-    }
-}
-
-void s19b_spark2_m_800D91DC(Work *work, int mode)
-{
-    if (mode < 8 && (work->f900 & 0x10))
-    {
-        work->f8EC = (void *)s19b_spark2_m_800D9148;
-        work->f8F4 = 0;
-        work->ctrl.turn.vz = 0;
-        work->ctrl.turn.vx = 0;
-        return;
-    }
-
-    if (mode >= 8)
-    {
-        work->f8E4->class |= 0x14;
-    }
-
-    if (s19b_spark2_m_800D88D8(work) != 0)
-    {
-        return;
-    }
-
-    if (work->f93C != 0 && mode >= 0xB)
-    {
-        work->f8EC = (void *)s19b_spark2_m_800D8B54;
-        work->f8F4 = 0;
-        work->ctrl.turn.vz = 0;
-        work->ctrl.turn.vx = 0;
-        work->f93C = 0;
-    }
-
-    if (mode == 0)
-    {
-        work->f8FC = 0x12;
-        GM_ConfigObjectAction((OBJECT *)&work->obj, 0x12, 0, 4);
-    }
-
-    if (work->obj.is_end == 0)
-    {
-        return;
-    }
-
-    work->f8EC = (void *)s19b_spark2_m_800D8B54;
-    work->f8F4 = 0;
-    work->ctrl.turn.vz = 0;
-    work->ctrl.turn.vx = 0;
-}
-
-void s19b_spark2_m_800D932C(Work *work, int mode);
-
-void s19b_spark2_m_800D92C8(Work *work, int mode)
-{
-    if (mode == 0)
-    {
-        work->f8FC = 10;
-        GM_ConfigObjectAction((OBJECT *)&work->obj, 10, 0, 4);
-    }
-    if (work->obj.is_end != 0)
-    {
-        work->f8EC = (void *)s19b_spark2_m_800D932C;
-        work->f8F4 = 0;
-        work->ctrl.turn.vz = 0;
-        work->ctrl.turn.vx = 0;
-    }
-}
-
-void s19b_spark2_m_800D932C(Work *work, int mode)
-{
-    if (mode == 0)
-    {
-        work->f8FC = 10;
-        GM_ConfigObjectAction((OBJECT *)&work->obj, 10, 0, 4);
-    }
-    if (work->obj.is_end != 0)
-    {
-        work->f8EC = (void *)s19b_spark2_m_800D932C;
-        work->f8F4 = 0;
-        work->ctrl.turn.vz = 0;
-        work->ctrl.turn.vx = 0;
-    }
-}
-
-void s19b_spark2_m_800D9390(Work *work, int mode)
-{
-    if (mode == 0)
-    {
-        work->f8FC = 20;
-        GM_ConfigObjectAction((OBJECT *)&work->obj, 20, 0, 4);
-        GM_SeSet(&work->ctrl.mov, 0x81);
-        s19b_spark2_m_800D8724(work, 5, 1);
-        *work->f944 |= 1;
-    }
-    if (work->obj.is_end != 0)
-    {
-        work->f8EC = (void *)s19b_spark2_m_800D8B54;
-        work->f8F4 = 0;
-        work->ctrl.turn.vz = 0;
-        work->ctrl.turn.vx = 0;
-        work->f940 = 45;
-        work->f93C = 0;
-    }
-}
-
-void s19b_spark2_m_800D9434(Work *work, int mode)
-{
-    if (s19b_spark2_m_800D899C(work) != 0)
-    {
-        return;
-    }
-    if (mode == 0)
-    {
-        GM_SeSet(&work->ctrl.mov, 0x81);
-        s19b_spark2_m_800D8724(work, 5, 1);
-        *work->f944 |= 1;
-    }
-    if (mode < 0x11)
-    {
-        return;
-    }
-    work->f8EC = (void *)s19b_spark2_m_800D8B54;
-    work->f8F4 = 0;
-    work->ctrl.turn.vz = 0;
-    work->ctrl.turn.vx = 0;
-}
-
-void s19b_spark2_m_800D94C8(Work *work, int mode)
-{
-    work->f8E4->class |= 0x14;
-    if (s19b_spark2_m_800D88D8(work) != 0)
-    {
-        return;
-    }
-    if (mode == 0)
-    {
-        work->f8FC = 0x13;
-        GM_ConfigObjectAction((OBJECT *)&work->obj, 0x13, 0, 4);
-    }
-    if (work->obj.is_end != 0)
-    {
-        work->f8EC = (void *)s19b_spark2_m_800D8B54;
-        work->f8F4 = 0;
-        work->ctrl.turn.vz = 0;
-        work->ctrl.turn.vx = 0;
-    }
-}
-
-void s19b_spark2_m_800D9558(Work *work)
-{
-    int   old;
-    void *handler;
-
-    work->f8E4->class = 1;
-    old     = work->f8F4;
-    handler = work->f8EC;
-    work->f8F4 = old + 1;
-
-    if (handler == NULL)
-    {
-        s19b_spark2_m_800D8918(work);
-        handler       = (void *)s19b_spark2_m_800D8B54;
-        work->f8EC    = handler;
-    }
-    ((void (*)(Work *, int))handler)(work, old);
-
-    s19b_spark2_m_800D8A48(work);
-
-    if (work->f940 != 0)
-    {
-        work->f8E4->class = 1;
-        work->f940 -= 1;
-    }
-}
-
-void s19b_spark2_m_800D95FC(Work *work)
-{
-    SVECTOR diff;
-
-    GV_SubVec3(&GM_PlayerPosition, &work->ctrl.mov, &diff);
-    diff.vy = 0;
-    work->f930 = GV_VecDir2(&diff);
-    work->f934 = GV_VecLen3(&diff);
-}
-
-extern void s19b_spark2_m_800D9C04(Work *work);
-
-void s19b_spark2_m_800D964C(Work *work)
-{
-    s19b_spark2_m_800D95FC(work);
-    s19b_spark2_m_800D9C04(work);
-    s19b_spark2_m_800D9558(work);
-}
-
-int s19b_spark2_m_800D9680(Work *work)
-{
-    int js = Takabe_JeepSystem.field_54;
-
-    if (js & 0x1000)
-    {
-        work->f914 = 4;
-        work->f918 = 0;
-        work->f900 |= 0x10;
-        return 1;
-    }
-    if (js & 0x2)
-    {
-        work->f900 |= 0x02000000;
-    }
-    else if (js & 0x8000)
-    {
-        work->f900 = 0x04000000;
-        return 1;
-    }
-    else if (work->f8FC == 0x14)
-    {
-        work->f914 = 4;
-        work->f918 = 0;
-        work->f900 |= 0x10;
-        return 1;
-    }
-    return 0;
-}
-
-void s19b_spark2_m_800D9704(Work *work)
-{
-    if (s19b_spark2_m_800D9680(work) != 0)
-    {
-        return;
-    }
-
-    if (work->f934 < 0xFA0 || work->f93C != 0)
-    {
-        int f930 = work->f930;
-        int new_f914;
-
-        if ((unsigned)(f930 - 0x601) < 0x3FF)
-        {
-            new_f914 = 1;
-        }
-        else if ((unsigned)(f930 - 0x201) < 0x3FF)
-        {
-            new_f914 = 3;
-        }
-        else if ((unsigned)(f930 - 0xA01) < 0x3FF)
-        {
-            new_f914 = 2;
-        }
-        else
-        {
-            work->f918 += 1;
-            return;
-        }
-
-        work->f914 = new_f914;
-        work->f918 = 0;
-        return;
-    }
-
-    work->f918 += 1;
-}
-
-void s19b_spark2_m_800D97A8(Work *work)
-{
-    int v;
-
-    if (s19b_spark2_m_800D9680(work) != 0)
-    {
-        return;
-    }
-
-    if (work->f8FC != 1)
-    {
-        work->f918 = 0;
-    }
-
-    if (work->f918 >= 0x4C)
-    {
-        work->f914 = 0;
-        work->f918 = 0;
-        return;
-    }
-
-    v = s19b_dword_800DE650;
-    if (work->f918 == v + 6 || work->f918 == v + 9 || work->f918 == v + 0xC)
-    {
-        work->f900 |= 8;
-    }
-    work->f900 |= 1;
-    work->f904 = work->f930;
-    work->f918 += 1;
-}
-
-void s19b_spark2_m_800D985C(Work *work)
-{
-    int v;
-
-    if (s19b_spark2_m_800D9680(work) != 0)
-    {
-        return;
-    }
-
-    if (work->f8FC != 3)
-    {
-        work->f918 = 0;
-    }
-
-    if (work->f918 >= 0x4C)
-    {
-        work->f914 = 0;
-        work->f918 = 0;
-        return;
-    }
-
-    v = s19b_dword_800DE650;
-    if (work->f918 == v + 6 || work->f918 == v + 9 || work->f918 == v + 0xC)
-    {
-        work->f900 |= 8;
-    }
-    work->f900 |= 2;
-    work->f904 = work->f930;
-    work->f918 += 1;
-}
-
-void s19b_spark2_m_800D9910(Work *work)
-{
-    int v;
-
-    if (s19b_spark2_m_800D9680(work) != 0)
-    {
-        return;
-    }
-
-    if (work->f8FC != 2)
-    {
-        work->f918 = 0;
-    }
-
-    if (work->f918 >= 0x4C)
-    {
-        work->f914 = 0;
-        work->f918 = 0;
-        return;
-    }
-
-    v = s19b_dword_800DE650;
-    if (work->f918 == v + 6 || work->f918 == v + 9 || work->f918 == v + 0xC)
-    {
-        work->f900 |= 8;
-    }
-    work->f900 |= 4;
-    work->f904 = work->f930;
-    work->f918 += 1;
-}
-
-void s19b_spark2_m_800D99C4(Work *work)
-{
-    int js = Takabe_JeepSystem.field_54;
-    work->f900 |= 0x10;
-    if (work->f918 >= 61 || (js & 0x2000))
-    {
-        work->f93C = 1;
-        work->f914 = 0;
-    }
-    if (js & 0x2)
-    {
-        work->f914 = 1;
-        work->f900 |= 0x02000000;
-    }
-    work->f918 += 1;
-}
-
-void s19b_spark2_m_800D9A30(Work *work)
-{
-    if (Takabe_JeepSystem.field_54 & 0x1000)
-    {
-        work->f914 = 4;
-        work->f918 = 0;
-        work->f900 |= 0x10;
-    }
-    else
-    {
-        work->f918 += 1;
-    }
-}
-
-void s19b_spark2_m_800D9A74(Work *work)
-{
-    int new_f900 = work->f900 | 0x10;
-    int js = Takabe_JeepSystem.field_54;
-    work->f900 = new_f900;
-    if (js & 0x2000)
-    {
-        work->f914 = 0;
-    }
-    work->f918 += 1;
-}
-
-#pragma INCLUDE_ASM("asm/overlays/s19b/s19b_spark2_m_800D9AA8.s")
-extern void s19b_spark2_m_800D9AA8(Work *work);
-
-void s19b_spark2_m_800D9B38(Work *work)
-{
-    switch (work->f914)
-    {
-    case 0:
-        s19b_spark2_m_800D9A30(work);
-        break;
-    case 4:
-        s19b_spark2_m_800D9A74(work);
-        break;
-    }
-}
-
-void s19b_spark2_m_800D9B88(Work *work)
-{
-    switch (work->f910)
-    {
-    case 0: s19b_spark2_m_800D9AA8(work); break;
-    case 1: s19b_spark2_m_800D9B38(work); break;
-    }
-
-    if (Takabe_JeepSystem.field_54 & 0x10000000)
-    {
-        work->f910 = 1;
-    }
-    else
-    {
-        work->f910 = 0;
-    }
-}
-
-void s19b_spark2_m_800D9C04(Work *work)
-{
-    short *table = (short *)&s19b_dword_800C3AC8;
-    int    idx;
-
-    work->f904 = -1;
-    work->f900 = 0;
-    idx = GM_GameLevel + 1;
-    s19b_dword_800DE650 = table[idx];
-
-    if (work->f90C == 0)
-    {
-        s19b_spark2_m_800D9B88(work);
-    }
-
-    Takabe_JeepSystem.field_54 &= 0xFFFF0000;
-    work->f920 += 1;
-}
-
-#pragma INCLUDE_ASM("asm/overlays/s19b/s19b_spark2_m_800D9C90.s")
-#pragma INCLUDE_ASM("asm/overlays/s19b/s19b_spark2_m_800D9EC0.s")
-void s19b_spark2_m_800DA0B4(POLY_FT4 *poly, int arg1, SVECTOR *from, SVECTOR *to)
-{
-    int d = (arg1 << 12) / 1178;
-    int dy;
-    int step;
-
-    poly->x0 += d;
-    poly->x1 += d;
-    poly->x2 += d;
-    poly->x3 += d;
-
-    dy = from->vy - to->vy;
-    if (arg1 < 0)
-    {
-        dy = -dy;
-    }
-    step = dy * 4;
-    poly->y2 = GV_NearExp4P(poly->y2, step);
-    poly->y3 = GV_NearExp4P(poly->y3, step);
-    *to = *from;
 }
